Add test for ConnectionsIteratorLazy
