+++ src/demo.test.ts
@@ -4,7 +4,7 @@ import runDemo from "./demo";
 test("demo", async () => {
   jest.setTimeout(90000);
 
-  const result = await runDemo();
+  const result = await runDemo(false);
 
   expect(result).toEqual(true);
 });
