+++ src/test/test-ldfetch.ts
@@ -4,7 +4,7 @@ import Rdf from "../util/Rdf";
 const ldfetch = new LDFetch({ headers: { Accept: "application/ld+json" } });
 
 // ldfetch.get("http://irail.be/stations/NMBS/008896008")
-ldfetch.get("https://belgium.linkedconnections.org/delijn/Oost-Vlaanderen/stops")
+ldfetch.get("https://openplanner.ilabt.imec.be/delijn/Limburg/stops")
   .then((response) => {
     Rdf.logTripleTable(response.triples);
   })
